+++ gulpfile.js/index.js
@@ -8,9 +8,10 @@ module.exports = {
   watchPreprocess: require('./tasks/watch-preprocess'),
   theoToMDX: require('./tasks/theo-to-mdx'),
   svgToTwig: require('./tasks/svg-to-twig'),
-  build: require('./tasks/build'),
   buildSass: require('./tasks/build-sass'),
-  buildJS: require('./tasks/build-js'),
-  buildTypes: require('./tasks/build-types'),
-  clean: require('./tasks/clean'),
+  buildJS: require('./tasks/build-scripts').buildJS,
+  // BuildTypes: require('./tasks/build-scripts').buildTypes,
+  // This is currently an empty function, because the buildTypes task fails if there are no types to build
+  // Once we add TS files with types, the empty function should be removed, and the above line should be un-commented
+  buildTypes: (done) => done(),
 };
