+++ package.json
@@ -106,7 +106,11 @@
     "preprocess:svg-to-twig": "gulp svgToTwig",
     "preprocess:theo-to-mdx": "gulp theoToMDX",
     "watch-preprocess": "gulp watchPreprocess",
-    "build": "gulp build",
+    "build:js": "gulp buildJS",
+    "build:types": "gulp buildTypes",
+    "build:sass": "gulp buildSass",
+    "build": "npm run clean && run-p build:*",
+    "clean": "eliminate dist",
     "type": "tsc",
     "type:watch": "tsc --watch",
     "check-lint": "run-s check-lint:*",
